api/requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-filter==23.5

# Database
psycopg2-binary==2.9.9
django-environ==0.11.2

# Cache and Session
django-redis==5.4.0
redis==5.0.1

# Search
meilisearch==0.31.0

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Celery for background tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# File handling and utilities
Pillow==10.1.0
python-decouple==3.8
PyMuPDF==1.26.5

# Development and debugging
django-extensions==3.2.3

# Production server
gunicorn==21.2.0